# api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: spond@temple.edu
    url: http://github.com/veg/api-datamonkey
  description: |
    Datamonkey is a free public server for comparative analysis of sequence alignments using state-of-the-art statistical models. <br> This is the OpenAPI definition for the Datamonkey API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Datamonkey API
  version: 1.0.0
  x-logo:
    url: https://raw.githubusercontent.com/veg/datamonkey-js/refs/heads/master/public/assets/img/header-logo.svg
    altText: Datamonkey.org logo
...
servers:
- url: http://datamonkey.org/api/v1
- url: https://datamonkey.org/api/v1
...
paths:
...
  /methods/gard-start:
    post:
      operationId: startGARDJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GardRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Success
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidDataError'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Start a GARD analysis
      tags:
      - GARD
  /methods/gard-result:
    post:
      operationId: getGARDJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GardRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GardResult'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Get GARD analysis results
      tags:
      - GARD
  /methods/bgm-start:
    post:
      operationId: startBGMJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BgmRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Success
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidDataError'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Start and monitor a BGM job
      tags:
      - BGM
  /methods/bgm-result:
    post:
      operationId: getBGMJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BgmRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BgmResult'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Get a BGM job result
      tags:
      - BGM
...
  /methods/fubar-start:
    post:
      description: |
        Starts a Fast Unconstrained Bayesian AppRoximation (FUBAR) job to detect sites under positive selection.
        FUBAR uses a Bayesian approach to infer selection pressures at individual sites, providing posterior
        probabilities of positive and negative selection.
      operationId: startFUBARJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FubarRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
          description: Success
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidDataError'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Start a FUBAR analysis
      tags:
      - FUBAR
  /methods/fubar-result:
    post:
      description: |
        Retrieves the results of a previously submitted Fast Unconstrained Bayesian AppRoximation (FUBAR) job.
        FUBAR uses a Bayesian approach to infer selection pressures at individual sites, providing posterior
        probabilities of positive and negative selection.
      operationId: getFUBARJob
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FubarRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FubarResult'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
          description: Internal Server Error
      summary: Get FUBAR analysis results
      tags:
      - FUBAR
...
components:
  schemas:
...
    GeneticCode:
      default: Universal
      description: |
        Available Genetic Codes
      enum:
      - Universal
      - Vertebrate mtDNA
      - Yeast mtDNA
      - Mold/Protozoan mtDNA
      - Invertebrate mtDNA
      - Ciliate Nuclear
      - Echinoderm mtDNA
      - Eupltoid Nuclear
      - Alt. Yeast Nuclear
      - Ascidian mtDNA
      - Flatworm mtDNA
      - Blepharisma Nuclear
      type: string
...
    JobStatus:
      example:
        error_message: error_message
        job_id: job_id
        status: queued
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        status:
          enum:
          - queued
          - running
          - completed
          - error
          type: string
        error_message:
          type: string
      type: object
    InvalidDataError:
      example:
        errors:
        - field: field
          message: message
        - field: field
          message: message
      properties:
        errors:
          items:
            $ref: '#/components/schemas/InvalidDataError_errors_inner'
          type: array
      type: object
    ServerError:
      example:
        code: code
        details: details
        error: error
      properties:
        error:
          description: A general description of the error.
          type: string
        code:
          description: A unique error code (if available).
          type: string
        details:
          description: Optional details about the error.
          type: string
      type: object
...
    BaseResponse:
      properties:
        analysis:
          $ref: '#/components/schemas/AnalysisInfo'
        input:
          $ref: '#/components/schemas/InputInfo'
        tested:
          additionalProperties:
            additionalProperties:
              enum:
              - tested
              - background
              type: string
            type: object
          type: object
        timers:
          additionalProperties:
            $ref: '#/components/schemas/TimersInfo_value'
          type: object
        runtime:
          pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+$"
          type: string
        data partitions:
          additionalProperties:
            $ref: '#/components/schemas/PartitionsInfo_value'
          type: object
      type: object
...
    GardRequest:
      example:
        genetic_code: Universal
        run_mode: Normal
        data_type: Nucleotide
        model: JTT
        alignment: alignment
        site_to_site_variation: None
        rate_classes: 0
      properties:
        alignment:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        genetic_code:
          $ref: '#/components/schemas/GeneticCode'
        data_type:
          default: Nucleotide
          description: The type of data being analyzed
          enum:
          - Nucleotide
          - Protein
          type: string
        run_mode:
          default: Normal
          description: The optimization mode
          enum:
          - Normal
          - Faster
          type: string
        site_to_site_variation:
          default: None
          description: Specifies the model for rate variation among sites
          enum:
          - None
          - General Discrete
          - Beta-Gamma
          type: string
        rate_classes:
          default: 2
          description: The number of discrete rate classes for rate variation
          type: integer
        model:
          default: JTT
          description: The substitution model to use
          type: string
      type: object
    GardResult:
      example:
        result:
          timers:
            key:
              timer: 6
              order: 0
          baselineScore: 0.8008281904610115
          tested:
            key:
              key: tested
          siteBreakPointSupport:
            key: 3.616076749251911
          runtime: runtime
          singleTreeAICc: 1.4658129805029452
          analysis:
            settings:
              site_filter: "{}"
            contact: john.smith@example.com
          trees:
            key: "{}"
          timeElapsed: 5
          input:
            trees:
              key: trees
          data partitions:
            key:
              coverage:
              - - 1
                - 1
              - - 1
                - 1
              name: name
          breakpointData:
            key:
              bps:
              - - 2
                - 2
              - - 2
                - 2
              tree: tree
          totalModelCount: 5
          potentialBreakpoints: 6
          improvements:
            key:
              deltaAICc: 9.301444243932576
              breakpoints:
              - - 7
                - 7
              - - 7
                - 7
        job_id: job_id
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        result:
          $ref: '#/components/schemas/GardResult_result'
      type: object
    BgmRequest:
      example:
        genetic_code: Universal
        max_parents: 5
        min_subs: 5
        burn_in: 6
        tree: tree
        data_type: codon
        alignment: alignment
        branches: All
        steps: 0
        samples: 1
      properties:
        alignment:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        tree:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        branches:
          default: All
          description: Specify branches to test
          type: string
        genetic_code:
          $ref: '#/components/schemas/GeneticCode'
        data_type:
          default: codon
          description: The type of data being analyzed
          enum:
          - nucleotide
          - amino-acid
          - codon
          type: string
        steps:
          default: 100000
          description: Number of MCMC steps to sample
          type: integer
        burn_in:
          default: 10000
          description: Number of MCMC steps to discard as burn-in
          type: integer
        samples:
          default: 100
          description: Number of samples to extract from the chain
          type: integer
        max_parents:
          default: 1
          description: Maximum number of parents allowed per node
          type: integer
        min_subs:
          default: 1
          description: Minimum number of substitutions per site to include in the
            analysis
          type: integer
      type: object
    BgmResult:
      example:
        result:
          input:
            trees:
              key: trees
          timers:
            key:
              timer: 6
              order: 0
          data partitions:
            key:
              coverage:
              - - 1
                - 1
              - - 1
                - 1
              name: name
          branch attributes:
            key:
              key:
                key: 2.3021358869347655
          tested:
            key:
              key: tested
          MLE:
            headers:
            - - headers
              - headers
            - - headers
              - headers
            content:
            - - 5.637376656633329
              - 5.637376656633329
            - - 5.637376656633329
              - 5.637376656633329
          runtime: runtime
          analysis:
            settings:
              site_filter: "{}"
            contact: john.smith@example.com
          fits:
            key:
              Log Likelihood: 0.8008281904610115
              estimated parameters: 6
              Rate Distributions: "{}"
              AIC-c: 1.4658129805029452
              display order: 5
        job_id: job_id
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        result:
          $ref: '#/components/schemas/BgmResult_result'
      type: object
...
    FubarRequest:
      example:
        genetic_code: Universal
        grid_points: 8
        tree: tree
        alignment: alignment
        concentration_parameter: 0.6031428726887332
      properties:
        alignment:
          description: "Path to the in-frame codon alignment file (e.g., .fasta, .nex)"
          type: string
        tree:
          description: Path to the phylogenetic tree file
          type: string
        genetic_code:
          default: Universal
          description: Which genetic code should be used
          type: string
        grid_points:
          default: 20
          description: Number of grid points for the Bayesian analysis (must be between
            5 and 50)
          maximum: 50
          minimum: 5
          type: integer
        concentration_parameter:
          default: 0.5
          description: Concentration parameter for the Dirichlet prior in the Bayesian
            estimation
          maximum: 1
          minimum: 0.001
          type: number
      required:
      - alignment
      - tree
      type: object
    FubarResult:
      example:
        result:
          input:
            partition_count: 1
            file_name: file_name
            number_of_sequences: 0
            trees:
              key: trees
            number_of_sites: 6
          timers:
            key: "{}"
          MLE:
            headers:
            - - headers
              - headers
            - - headers
              - headers
            content:
              key:
              - - 5.962133916683182
                - 5.962133916683182
              - - 5.962133916683182
                - 5.962133916683182
          analysis:
            requirements: requirements
            citation: citation
            contact: contact
            version: version
            info: info
            authors: authors
          fits:
            key: "{}"
        job_id: job_id
      properties:
        job_id:
          pattern: "^[a-zA-Z0-9]+$"
          type: string
        result:
          $ref: '#/components/schemas/FubarResult_result'
      type: object
...
    InvalidDataError_errors_inner:
      example:
        field: field
        message: message
      properties:
        field:
          description: The field that caused the error
          type: string
        message:
          description: The error message
          type: string
      type: object
...
    TimersInfo_value:
      example:
        timer: 6
        order: 0
      properties:
        order:
          type: integer
        timer:
          type: integer
      type: object
...
    GardResult_result_allOf_breakpointData_value:
      example:
        bps:
        - - 2
          - 2
        - - 2
          - 2
        tree: tree
      properties:
        bps:
          description: Breakpoint positions
          items:
            items:
              type: integer
            type: array
          type: array
        tree:
          description: Newick tree for this segment
          type: string
      type: object
    GardResult_result_allOf_improvements_value:
      example:
        deltaAICc: 9.301444243932576
        breakpoints:
        - - 7
          - 7
        - - 7
          - 7
      properties:
        breakpoints:
          description: Breakpoint positions
          items:
            items:
              type: integer
            type: array
          type: array
        deltaAICc:
          description: Change in AICc score
          type: number
      type: object
    GardResult_result:
      allOf:
      - $ref: '#/components/schemas/BaseResponse'
      - properties:
          baselineScore:
            description: Baseline score for the analysis
            type: number
          potentialBreakpoints:
            description: Number of potential breakpoints analyzed
            type: integer
          singleTreeAICc:
            description: AICc score for the single tree model
            type: number
          timeElapsed:
            description: Time elapsed for the analysis in seconds
            type: integer
          totalModelCount:
            description: Total number of models evaluated
            type: integer
          breakpointData:
            additionalProperties:
              $ref: '#/components/schemas/GardResult_result_allOf_breakpointData_value'
            description: Data for each segment between breakpoints
            type: object
          improvements:
            additionalProperties:
              $ref: '#/components/schemas/GardResult_result_allOf_improvements_value'
            description: Improvement metrics for each breakpoint
            type: object
          siteBreakPointSupport:
            additionalProperties:
              type: number
            description: Support values for potential breakpoints at each site
            type: object
          trees:
            additionalProperties:
              type: object
            description: Tree data
            type: object
        type: object
      example:
        timers:
          key:
            timer: 6
            order: 0
        baselineScore: 0.8008281904610115
        tested:
          key:
            key: tested
        siteBreakPointSupport:
          key: 3.616076749251911
        runtime: runtime
        singleTreeAICc: 1.4658129805029452
        analysis:
          settings:
            site_filter: "{}"
          contact: john.smith@example.com
        trees:
          key: "{}"
        timeElapsed: 5
        input:
          trees:
            key: trees
        data partitions:
          key:
            coverage:
            - - 1
              - 1
            - - 1
              - 1
            name: name
        breakpointData:
          key:
            bps:
            - - 2
              - 2
            - - 2
              - 2
            tree: tree
        totalModelCount: 5
        potentialBreakpoints: 6
        improvements:
          key:
            deltaAICc: 9.301444243932576
            breakpoints:
            - - 7
              - 7
            - - 7
              - 7
...
    BgmResult_result:
      allOf:
      - $ref: '#/components/schemas/BaseResponse'
      - properties:
          fits:
            additionalProperties:
              $ref: '#/components/schemas/BgmResult_result_allOf_fits_value'
            description: Model fit information
            type: object
          MLE:
            $ref: '#/components/schemas/BgmResult_result_allOf_MLE'
          branch attributes:
            additionalProperties:
              additionalProperties:
                additionalProperties:
                  type: number
                type: object
              type: object
            description: Branch-specific attributes
            type: object
        type: object
      example:
        input:
          trees:
            key: trees
        timers:
          key:
            timer: 6
            order: 0
        data partitions:
          key:
            coverage:
            - - 1
              - 1
            - - 1
              - 1
            name: name
        branch attributes:
          key:
            key:
              key: 2.3021358869347655
        tested:
          key:
            key: tested
        MLE:
          headers:
          - - headers
            - headers
          - - headers
            - headers
          content:
          - - 5.637376656633329
            - 5.637376656633329
          - - 5.637376656633329
            - 5.637376656633329
        runtime: runtime
        analysis:
          settings:
            site_filter: "{}"
          contact: john.smith@example.com
        fits:
          key:
            Log Likelihood: 0.8008281904610115
            estimated parameters: 6
            Rate Distributions: "{}"
            AIC-c: 1.4658129805029452
            display order: 5
...
    FubarResult_result:
      example:
        input:
          partition_count: 1
          file_name: file_name
          number_of_sequences: 0
          trees:
            key: trees
          number_of_sites: 6
        timers:
          key: "{}"
        MLE:
          headers:
          - - headers
            - headers
          - - headers
            - headers
          content:
            key:
            - - 5.962133916683182
              - 5.962133916683182
            - - 5.962133916683182
              - 5.962133916683182
        analysis:
          requirements: requirements
          citation: citation
          contact: contact
          version: version
          info: info
          authors: authors
        fits:
          key: "{}"
      properties:
        analysis:
          $ref: '#/components/schemas/FubarResult_result_analysis'
        input:
          $ref: '#/components/schemas/FubarResult_result_input'
        fits:
          additionalProperties:
            type: object
          description: Model fit statistics
          type: object
        timers:
          additionalProperties:
            type: object
          description: Timing information for different stages of the analysis
          type: object
        MLE:
          $ref: '#/components/schemas/ContrastFelResult_result_MLE'
      type: object
...